Rename the rng folder to fiat_shamir
