+++ marlin/src/fiat_shamir/traits/algebraic_sponge.rs
