+++ marlin/src/fiat_shamir/traits/fiat_shamir.rs
